refactor(client): extract product tests to sibling file
